the pain in rain falls mainly on the gutter
hhhhh aaaaaaaa okay yeah I'm glad I'm using git because I nearly
screwed myself over.

+++ labs/5.22/LoanPayment.java
@@ -1,6 +1,6 @@
 public abstract class LoanPayment {
-   public short paymentNumber;
-   public short paymentsRemaining;
+   public int paymentNumber;
+   public int paymentsRemaining;
 
    public double interestPaid;
    public double principalPaid;
